requirements.txt
SpeechRecognition==3.10.3
openai==1.25.1
sounddevice==0.4.6
soundfile==0.12.1
deepgram-sdk==3.2.7
python-dotenv==1.0.1
pyaudio==0.2.14